Rename import files and classes based on these formats: - Format: Change (original) obs_instrument_{name} to (new)   obs_volume_{name}, and for pds4: obs_bundle_{name} Format: Change (original) obs_mission_{name} to (new)   obs_volume_{name}_helper, and for pds4: obs_bundle_{name}_helper
