+++ opus/import/obs_volume_couvis.py
@@ -1,7 +1,7 @@
 ################################################################################
-# obs_instrument_couvis.py
+# obs_volume_couvis.py
 #
-# Defines the ObsInstrumentCOUVIS class, which encapsulates fields in the
+# Defines the ObsVolumeCOUVIS class, which encapsulates fields in the
 # common, obs_mission_cassini, and obs_instrument_couvis tables for COUVIS_0xxx.
 ################################################################################
 
@@ -9,10 +9,10 @@ import os
 
 import opus_support
 
-from obs_mission_cassini import ObsMissionCassini
+from obs_volume_cassini_helper import ObsVolumeCassiniHelper
 
 
-class ObsInstrumentCOUVIS(ObsMissionCassini):
+class ObsVolumeCOUVIS(ObsVolumeCassiniHelper):
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
 
@@ -331,7 +331,7 @@ class ObsInstrumentCOUVIS(ObsMissionCassini):
 
 
     #######################################
-    ### OVERRIDE FROM ObsMissionCassini ###
+    ### OVERRIDE FROM ObsVolumeCassiniHelper ###
     #######################################
 
     def field_obs_mission_cassini_spacecraft_clock_count1(self):
