+++ opus/import/obs_volume_couvis_occ.py
@@ -1,17 +1,17 @@
 ################################################################################
-# obs_instrument_couvis_occ.py
+# obs_volume_couvis_occ.py
 #
-# Defines the ObsInstrumentCOUVISOcc class, which encapsulates fields in the
+# Defines the ObsVolumeCOUVISOcc class, which encapsulates fields in the
 # common, obs_mission_cassini, and obs_instrument_couvis tables for COUVIS_8001
 # occultations.
 ################################################################################
 
 import opus_support
 
-from obs_instrument_couvis_covims_occ import ObsInstrumentUVISVIMSOcc
+from obs_volume_couvis_covims_occ import ObsVolumeUVISVIMSOcc
 
 
-class ObsInstrumentCOUVISOcc(ObsInstrumentUVISVIMSOcc):
+class ObsVolumeCOUVISOcc(ObsVolumeUVISVIMSOcc):
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
 
@@ -60,7 +60,7 @@ class ObsInstrumentCOUVISOcc(ObsInstrumentUVISVIMSOcc):
 
 
     #######################################
-    ### OVERRIDE FROM ObsMissionCassini ###
+    ### OVERRIDE FROM ObsVolumeCassiniHelper ###
     #######################################
 
     def field_obs_mission_cassini_spacecraft_clock_count1(self):
