+++ opus/import/obs_volume_ebrocc.py
@@ -1,7 +1,7 @@
 ################################################################################
-# obs_instrument_ebrocc.py
+# obs_volume_ebrocc.py
 #
-# Defines the ObsInstrumentEBROCC class, which encapsulates fields for
+# Defines the ObsVolumeEBROCC class, which encapsulates fields for
 # the common tables for EBROCC_0001. This class supports multiple instruments
 # in a single volume.
 ################################################################################
@@ -17,7 +17,7 @@ from obs_common import ObsCommon
 # * Emission angle and north-based emission angle = incidence angle
 # * Observer elevation = 90 - incidence angle
 
-class ObsInstrumentEBROCC(ObsCommon):
+class ObsVolumeEBROCC(ObsCommon):
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
