+++ opus/import/obs_volume_hststis.py
@@ -1,16 +1,16 @@
 ################################################################################
-# obs_instrument_hststis.py
+# obs_volume_hststis.py
 #
-# Defines the ObsInstrumentHSTSTIS class, which encapsulates fields in the
+# Defines the ObsVolumeHSTSTIS class, which encapsulates fields in the
 # common and obs_mission_hubble tables for the HST STIS instrument for
 # HSTOx_xxxx. Note HST does not have separate tables for each instrument but
 # combines them all together.
 ################################################################################
 
-from obs_mission_hubble import ObsMissionHubble
+from obs_volume_hubble_helper import ObsVolumeHubbleHelper
 
 
-class ObsInstrumentHSTSTIS(ObsMissionHubble):
+class ObsVolumeHSTSTIS(ObsVolumeHubbleHelper):
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
 
@@ -108,9 +108,9 @@ class ObsInstrumentHSTSTIS(ObsMissionHubble):
         return self._create_mult('NONE')
 
 
-    ######################################
-    ### OVERRIDE FROM ObsMissionHubble ###
-    ######################################
+    ###########################################
+    ### OVERRIDE FROM ObsVolumeHubbleHelper ###
+    ###########################################
 
     def field_obs_mission_hubble_filter_type(self):
         filter1, filter2 = self._decode_filters()
